refactor(frontend): extract component for IC transactions scrolling
